Greek: Fix the uppercase theta translit mapping
transliterate python module has a wrong mapping for the Greek theta
letters. There is no mapping for the uppercase theta letter while there
is a mapping for the Greek theta symbol U+03D1. The small and capital
theta letters emitted by Greek keyboards are U+03B8 and U+0398
respectively. Remove the mapping for the symbol, fix the mapping for the
small theta letter, add the mapping for the capital theta letter.